+++ src/components/LineChart/constants.ts
@@ -7,3 +7,6 @@ export const SMALL_WIDTH = 300;
 export const MIN_LABEL_SPACE = 100;
 export const TICK_SIZE = 6;
 export const FONT_SIZE = 12;
+export const ANIMATION_DELAY = 0.25;
+export const SLOW_DURATION = 1;
+export const FAST_DURATION = 0.5;