+++ src/components/LineChart/tests/Chart.test.tsx
@@ -56,6 +56,7 @@ const mockProps = {
   renderTooltipContent: jest.fn(() => <p>Mock Tooltip</p>),
   hideXAxisLabels: false,
   hasSpline: false,
+  isAnimated: false,
 };
 
 const mockEmptyStateProps = {
@@ -67,11 +68,20 @@ const mockEmptyStateProps = {
   renderTooltipContent: jest.fn(() => <p>Mock Tooltip</p>),
   hideXAxisLabels: false,
   hasSpline: false,
+  isAnimated: false,
 };
 
 describe('<Chart />', () => {
   beforeEach(() => {
     jest.useFakeTimers();
+  });
+
+  beforeAll(() => {
+    Object.defineProperty(window, 'matchMedia', {
+      value: jest.fn(() => {
+        return {matches: false};
+      }),
+    });
   });
 
   afterEach(() => {